+++ packages/tsx-editor/src/transpiler/transpile.ts
@@ -1,7 +1,7 @@
 import * as monaco from 'monaco-editor';
 import { TypeScriptWorker, EmitOutput } from './monacoTypescriptWorker.d';
 import { transformExample } from './exampleTransform';
-import { _getErrorMessages } from './helpers';
+import { _getErrorMessages } from './transpileHelpers';
 
 export interface ITranspiledOutput {
   outputString?: string;
